extract out a point adapter, in preparation for adding unit tests
